Complete the round trip test for encode and sort bin mat - Updated comments for clarity in bin_mat_payload.cc. - Improved readability and consistency in genotype_coder.cc. - Added comments for function inputs and outputs in genotype_coder.h. - Fixed bit shift operations in genotype_parameters.cc. - Cleaned up unnecessary lines and improved spacing in genotype_payload.cc. - Removed duplicate includes in genotype_payload.h. - Changed vector types to uint32_t for consistency and efficiency in row_col_ids_payload.cc and row_col_ids_payload.h. - Improved size calculations and added assertions in sorted_bin_mat_payload.cc. - Added detailed test descriptions and checks in encode_test.cc to cover various sorting scenarios. - Improved readability and added more test cases in structure_test.cc. - Completed the round trip test for encode_and_sort_bin_mat in encode_test.cc.

+++ src/genie/genotype/genotype_payload.cc
@@ -310,9 +310,9 @@ void GenotypePayload::Write(util::BitWriter& writer) const {
   writer.WriteBypassBE(GetMaxPloidy());
 
   uint8_t flag = 0;
-  flag |= GetNoReferenceFlag() << 2u;
-  flag |= GetNotAvailableFlag() << 1u;
-  flag |= GetPhasesValue() << 0u;
+  flag |= static_cast<uint8_t>(GetNoReferenceFlag()) << 2u;
+  flag |= static_cast<uint8_t>(GetNotAvailableFlag()) << 1u;
+  flag |= static_cast<uint8_t>(GetPhasesValue()) << 0u;
   writer.WriteBypassBE(flag);
 
   writer.WriteBypassBE(GetNumBitPlanes());
@@ -336,9 +336,9 @@ void GenotypePayload::Write(core::Writer& writer) const {
   writer.Write(GetMaxPloidy(),8);
 
   uint8_t flag = 0;
-  flag |= GetNoReferenceFlag() << 2u;
-  flag |= GetNotAvailableFlag() << 1u;
-  flag |= GetPhasesValue() << 0u;
+  flag |= static_cast<uint8_t>(GetNoReferenceFlag()) << 2u;
+  flag |= static_cast<uint8_t>(GetNotAvailableFlag()) << 1u;
+  flag |= static_cast<uint8_t>(GetPhasesValue()) << 0u;
   writer.Write(flag,8);
 
   writer.Write(GetNumBitPlanes(),8);
